requirements.txt
Django>=4.2.0,<5.0.0
mysql-connector-python>=8.0.0
redis>=4.0.0
requests>=2.25.0
cryptography>=3.4.0
